fix: use merged schema with input data to check for two steps requirements instead
When schema is detected from data for dsv file, it may add additional columns that are set as ignored. Therefore, we then go into a two step ingestion in order to be able to add those additional columns before ignoring them. If not, then we may have a shift on column ingestion and have an error because of the column type or worse, data into the wrong column.